+++ packages/eslint-config/index.js
@@ -21,6 +21,18 @@ module.exports = {
   },
   rules: {
     'prettier/prettier': 'error',
+    'react/jsx-props-no-spreading': 'off',
+    'import/no-extraneous-dependencies': [
+      'error',
+      {
+        devDependencies: [
+          '**/*.test.*',
+          '**/*.test.*',
+          'test/**',
+          '__test__/*',
+        ],
+      },
+    ],
     'react/jsx-filename-extension': [
       'error',
       { extensions: ['.js', '.jsx', 'tsx', 'ts'] },