+++ packages/mango-ui/src/index.ts
@@ -1,14 +1,19 @@
 import Button from './components/Button';
-export * from './components/Button';
 
 import Editor from './components/Editor';
-export * from './components/Editor';
 
 import Input from './components/Input';
+
+import Radio from './components/Radio';
+
+export * from './components/Button';
+export * from './components/Editor';
 export * from './components/Input';
+export * from './components/Radio';
 
 export default {
   Button,
   Editor,
   Input,
+  Radio,
 };