+++ packages/stories/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.11](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.10...@mango-solution/stories@0.0.11) (2020-06-29)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
+## [0.0.10](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.9...@mango-solution/stories@0.0.10) (2020-06-27)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
+## [0.0.9](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.8...@mango-solution/stories@0.0.9) (2020-06-26)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
 ## [0.0.8](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.7...@mango-solution/stories@0.0.8) (2020-06-14)
 
 **Note:** Version bump only for package @mango-solution/stories